pyproject.toml
[tool.poetry]
name = "consumables_app"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.1"
Flask = "3.0.3"
sqlalchemy = "2.0.30"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
flask-pydantic = "^0.12.0"
gunicorn = "^22.0.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
